# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/molecules": {
      "post": {
        "operationId": "MoleculesController_create",
        "summary": "Create a new molecule",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMoleculeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Molecules"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "MoleculesController_findAll",
        "summary": "Get all molecules",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Molecules"
        ]
      }
    },
    "/api/v1/molecules/{id}": {
      "get": {
        "operationId": "MoleculesController_findOne",
        "summary": "Get a molecule by id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Molecules"
        ]
      },
      "patch": {
        "operationId": "MoleculesController_update",
        "summary": "Update a molecule",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMoleculeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Molecules"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "MoleculesController_remove",
        "summary": "Delete a molecule",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Molecules"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/molecules/{id}/propose-changes": {
      "post": {
        "operationId": "MoleculesController_proposeChanges",
        "summary": "Propose changes to a molecule",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMoleculeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Molecules"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/reactions": {
      "post": {
        "operationId": "ReactionsController_create",
        "summary": "Create a new reaction",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReactionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Reactions"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "ReactionsController_findAll",
        "summary": "Get all reactions",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Reactions"
        ]
      }
    },
    "/api/v1/reactions/{id}": {
      "get": {
        "operationId": "ReactionsController_findOne",
        "summary": "Get a reaction by id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Reactions"
        ]
      },
      "patch": {
        "operationId": "ReactionsController_update",
        "summary": "Update a reaction",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReactionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Reactions"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "ReactionsController_remove",
        "summary": "Delete a reaction",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Reactions"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/reactions/{id}/propose-changes": {
      "post": {
        "operationId": "ReactionsController_proposeChanges",
        "summary": "Propose changes to a reaction",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReactionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Reactions"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "♻️ Recycling Centers Management API",
    "description": "Enterprise-grade API for managing recycling centers, materials, and user operations",
    "version": "1.0",
    "contact": {
      "name": "The Wealthiest Programmer",
      "url": "https://github.com/wealthiest-programmer",
      "email": "contact@wealthiest-programmer.dev"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateMoleculeDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the molecule"
          },
          "description": {
            "type": "string",
            "description": "Description of the molecule"
          },
          "smiles": {
            "type": "string",
            "description": "SMILES string"
          },
          "inchi": {
            "type": "string",
            "description": "InChI string"
          },
          "formula": {
            "type": "string",
            "description": "Molecular formula"
          },
          "weight": {
            "type": "number",
            "description": "Molecular weight"
          },
          "structure": {
            "type": "object",
            "description": "Structure data (JSON)"
          },
          "properties": {
            "type": "object",
            "description": "Properties data (JSON)"
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateMoleculeDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the molecule"
          },
          "description": {
            "type": "string",
            "description": "Description of the molecule"
          },
          "smiles": {
            "type": "string",
            "description": "SMILES string"
          },
          "inchi": {
            "type": "string",
            "description": "InChI string"
          },
          "formula": {
            "type": "string",
            "description": "Molecular formula"
          },
          "weight": {
            "type": "number",
            "description": "Molecular weight"
          },
          "structure": {
            "type": "object",
            "description": "Structure data (JSON)"
          },
          "properties": {
            "type": "object",
            "description": "Properties data (JSON)"
          }
        },
        "required": [
          "name"
        ]
      },
...
      "CreateReactionDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the reaction"
          },
          "description": {
            "type": "string",
            "description": "Description"
          },
          "reactionString": {
            "type": "string",
            "description": "Reaction string (SMILES/SMARTS)"
          },
          "data": {
            "type": "object",
            "description": "ORD Data (JSON)"
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateReactionDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the reaction"
          },
          "description": {
            "type": "string",
            "description": "Description"
          },
          "reactionString": {
            "type": "string",
            "description": "Reaction string (SMILES/SMARTS)"
          },
          "data": {
            "type": "object",
            "description": "ORD Data (JSON)"
          }
        },
        "required": [
          "name"
        ]
      },
...
    }
  }
}